// src/app/modules/characters/external/interfaces/people.interfaces.ts
export interface GetPeopleResponse {
  count: number;
  next: string;
  previous: any;
  results: SwapiCharacter[];
}

export interface SwapiCharacter {
  name: string;
  height: string;
  mass: string;
  hair_color: string;
  skin_color: string;
  eye_color: string;
  birth_year: string;
  gender: string;
  homeworld: string;
  films: string[];
  species: string[];
  vehicles: string[];
  starships: string[];
  created: string;
  edited: string;
  url: string;
}